# Conflicts:
#	src/main/java/bean/pojo/Relation.java

+++ src/main/java/bean/pojo/Grade.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Grade {
-  private long gradeId;
+  private Integer gradeId;
   private String gradeName;
   private String gradeContent;
   private double gradeCharge;
